chore: restructure entities types
