+++ libs/entities/src/enums/activity/index.ts
