+++ libs/entities/src/enums/common/index.ts
