+++ libs/entities/src/enums/index.ts
@@ -1,1 +1,3 @@
-export * from './common.enum';
+export * from './common';
+export * from './activity';
+export * from './organization';
