+++ libs/entities/src/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
 export * from './enums';
-export * from './common';
