--- libs/entities/src/types/common.type.ts
@@ -1,16 +1,0 @@
-import { EPaginationOrderBy } from '../enums';
-import { TJwtRequest } from './auth.type';
-
-export type THeaderRequest = {
-  user: TJwtRequest & {
-    fullname?: string;
-    avatar?: string;
-  };
-};
-
-export type TPaginationRequest = {
-  page: string;
-  perPage: string;
-  orderBy: EPaginationOrderBy;
-  search: string;
-};
