+++ libs/entities/src/types/index.ts
@@ -1,4 +1,4 @@
-export * from './auth.type';
-export * from './user.type';
-export * from './common.type';
-export * from './activity.type';
+export * from './auth';
+export * from './user';
+export * from './common';
+export * from './activity';
